Fix blocking thread issue on the Webflux variant
Prior to this commit, the `ExampleService` would use `Thread.sleep`
calls to simulate some latency in the service. This is not the right
approach: the runtime model of Reactor is operating on very few threads
and an event loop. Calling `Thread.sleep` effectively blocks one of
those few threads and will have devastating effects on the performance.

There are two ways to simulate latency:

1. Wrap the blocking call on a different scheduler.
See https://projectreactor.io/docs/core/release/reference/#faq.wrap-blocking
This is what other frameworks featured in this benchmark are doing.

2. Use the `.delay*` operators to delay the `onNext` signal with the
expected value. This is the solution chosen by this PR as it is much
closer to the typical call made in reactive applications (i.e. a REST
client call, a database call, etc).